Add `TransactionFailedEvent`
