+++ src/backend/app/Jobs/EnergyTransactionProcessor.php
@@ -2,6 +2,7 @@
 
 namespace App\Jobs;
 
+use App\Events\TransactionFailedEvent;
 use App\Exceptions\TransactionAmountNotEnoughException;
 use App\Exceptions\TransactionNotInitializedException;
 use App\Misc\TransactionDataContainer;
@@ -40,7 +41,7 @@ class EnergyTransactionProcessor extends AbstractJob {
             }
         } catch (\Exception $e) {
             Log::info('Transaction failed.: '.$e->getMessage());
-            event('transaction.failed', [$this->transaction, $e->getMessage()]);
+            event(new TransactionFailedEvent($this->transaction, $e->getMessage()));
         }
     }
 
@@ -54,7 +55,7 @@ class EnergyTransactionProcessor extends AbstractJob {
         try {
             return TransactionDataContainer::initialize($this->transaction);
         } catch (\Exception $e) {
-            event('transaction.failed', [$this->transaction, $e->getMessage()]);
+            event(new TransactionFailedEvent($this->transaction, $e->getMessage()));
             throw new TransactionNotInitializedException($e->getMessage());
         }
     }
