+++ src/backend/database/seeders/TransactionSeeder.php
@@ -3,6 +3,7 @@
 namespace Database\Seeders;
 
 use App\Events\PaymentSuccessEvent;
+use App\Events\TransactionFailedEvent;
 use App\Models\Device;
 use App\Models\MainSettings;
 use App\Models\Meter\Meter;
@@ -241,7 +242,7 @@ class TransactionSeeder extends Seeder {
             // create an object for the token job
             $transactionData = \App\Misc\TransactionDataContainer::initialize($transaction);
         } catch (\Exception $exception) {
-            event('transaction.failed', [$transaction, $exception->getMessage()]);
+            event(new TransactionFailedEvent($transaction, $exception->getMessage()));
             throw $exception;
         }
