+++ src/backend/packages/inensus/mesomb-payment-provider/src/Listeners/TransactionSuccessListener.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Listeners;
+
+use App\Models\Transaction\Transaction;
+
+class TransactionSuccessListener {
+    public function onTransactionSuccess(Transaction $transaction) {
+        $transactionProvider = resolve('MesombPaymentProvider');
+        $transactionProvider->sendResult(true, $transaction);
+    }
+
+    public function handle(Transaction $transaction): void {
+        $this->onTransactionSuccess($transaction);
+    }
+}
