+++ src/backend/packages/inensus/mesomb-payment-provider/src/Providers/EventServiceProvider.php
@@ -2,12 +2,15 @@
 
 namespace Inensus\MesombPaymentProvider\Providers;
 
+use App\Events\TransactionFailedEvent;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-use Inensus\MesombPaymentProvider\Listeners\TransactionListener;
+use Inensus\MesombPaymentProvider\Listeners\TransactionFailedListener;
+use Inensus\MesombPaymentProvider\Listeners\TransactionSuccessfulListener;
 
 class EventServiceProvider extends ServiceProvider {
     protected $subscribe = [
-        //   TransactionListener::class
+        // TransactionFailedEvent::class => [TransactionFailedListener::class],
+        // 'transaction.successful' => [TransactionSuccessfulListener::class],
     ];
 
     /**
